ML/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_api*", "relativa_ml*"]

[project]
name = "relativa-ml"
version = "0.1.0"
description = "Relativa ML microservice (Django)"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.1,<5.2",
    "djangorestframework>=3.15",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "scikit-learn>=1.5",
]

[project.optional-dependencies]
dev = ["black", "isort", "ruff"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]
